fix: (offtopic) add thumbnails to imported/migrated files
